chore(Typings): moved typings in src folder
