+++ package.json
@@ -4,7 +4,7 @@
   "description": "A low level blockchain explorers lookup",
   "main": "lib/cjs",
   "module": "lib/esm",
-  "typings": "./typings/index.d.ts",
+  "typings": "lib/index.d.ts",
   "dependencies": {
     "xmlhttprequest": "^1.8.0"
   },
@@ -28,7 +28,8 @@
     "typescript": "^4.1.3"
   },
   "scripts": {
-    "compile": "tsc -p tsconfig.json && tsc -p tsconfig.cjs.json",
+    "copy:declaration": "cp src/typings/index.d.ts lib/index.d.ts",
+    "compile": "tsc -p tsconfig.json && tsc -p tsconfig.cjs.json && npm run copy:declaration",
     "lint": "eslint . --ext .ts",
     "test": "jest",
     "test:watch": "jest --watch",
