+++ src/constants/api.ts
@@ -1,9 +1,1 @@
-export enum TRANSACTION_APIS {
-  bitpay = 'bitpay',
-  blockcypher = 'blockcypher',
-  blockexplorer = 'blockexplorer',
-  blockstream = 'blockstream',
-  etherscan = 'etherscan'
-}
-
 export const TRANSACTION_ID_PLACEHOLDER = '{transaction_id}';
