+++ src/explorers/ethereum/blockcypher.ts
@@ -1,8 +1,7 @@
 import { stripHashPrefix } from '../../utils/stripHashPrefix';
 import { BLOCKCHAINS } from '../../constants/blockchains';
-import { TransactionData } from '../../models/TransactionData';
-import { TRANSACTION_APIS, TRANSACTION_ID_PLACEHOLDER } from '../../constants/api';
-import { ExplorerAPI, ExplorerURLs, IParsingFunctionAPI } from '../../models/Explorers';
+import { ExplorerAPI, ExplorerURLs, IParsingFunctionAPI, TRANSACTION_APIS, TransactionData } from '../../typings';
+import { TRANSACTION_ID_PLACEHOLDER } from '../../constants/api';
 import CONFIG from '../../constants/config';
 import { dateToUnixTimestamp } from '../../utils/date';
 import { prependHashPrefix } from '../../utils/prependHashPrefix';
