+++ src/explorers/rpc/bitcoin.ts
@@ -1,8 +1,7 @@
-import { TransactionData } from '../../models/TransactionData';
 import { stripHashPrefix } from '../../utils/stripHashPrefix';
 import { timestampToDateObject } from '../../utils/date';
 import { request } from '../../services/request';
-import { IParsingFunctionAPI } from '../../models/Explorers';
+import { IParsingFunctionAPI, TransactionData } from '../../typings';
 
 export async function bitcoinRPCParsingFunction ({ serviceUrl, transactionId }: IParsingFunctionAPI): Promise<TransactionData> {
   const getRawTransactionParams = {
