--- src/models/TransactionData.ts
@@ -1,6 +1,0 @@
-export interface TransactionData {
-  remoteHash: string;
-  issuingAddress: string;
-  time: string | Date;
-  revokedAddresses: string[];
-}
