+++ src/typings/index.d.ts
@@ -1,8 +1,29 @@
-import { TRANSACTION_APIS } from '../constants/api';
-import { SupportedChains } from '../constants/blockchains';
-import { TransactionData } from './TransactionData';
+export declare interface TransactionData {
+  remoteHash: string;
+  issuingAddress: string;
+  time: string | Date;
+  revokedAddresses: string[];
+}
 
-export interface ExplorerURLs {
+export declare enum SupportedChains {
+  Bitcoin = 'bitcoin',
+  Ethmain = 'ethmain',
+  Ethropst = 'ethropst',
+  Ethrinkeby = 'ethrinkeby',
+  Mocknet = 'mocknet',
+  Regtest = 'regtest',
+  Testnet = 'testnet'
+}
+
+export declare enum TRANSACTION_APIS {
+  bitpay = 'bitpay',
+  blockcypher = 'blockcypher',
+  blockexplorer = 'blockexplorer',
+  blockstream = 'blockstream',
+  etherscan = 'etherscan'
+}
+
+export declare interface ExplorerURLs {
   main: string;
   test: string;
 }
@@ -11,7 +32,8 @@ export type TExplorerFunctionsArray = Array<{
   getTxData: (transactionId: string, chain?: SupportedChains) => Promise<TransactionData>;
   priority?: number;
 }>;
-export interface IParsingFunctionAPI {
+
+export declare interface IParsingFunctionAPI {
   jsonResponse?: any; // the response from the service when called as rest
   chain?: SupportedChains; // TODO: look at how to deprecate this. Only used in etherscan
   key?: string; // identification key to pass to the service -> TODO: can this be merged into the serviceUrl? Only used in etherscan
@@ -19,10 +41,10 @@ export interface IParsingFunctionAPI {
   transactionId?: string; // when using in RPCs we pass the tx id to look up since these functions are responsible for service lookup
   serviceUrl?: string; // the distant service url
 }
-export type TExplorerParsingFunction = ((data: IParsingFunctionAPI) => TransactionData) |
-((data: IParsingFunctionAPI) => Promise<TransactionData>);
 
-export interface ExplorerAPI {
+export type TExplorerParsingFunction = ((data: IParsingFunctionAPI) => TransactionData) | ((data: IParsingFunctionAPI) => Promise<TransactionData>);
+
+export declare interface ExplorerAPI {
   serviceURL?: string | ExplorerURLs;
   priority?: 0 | 1 | -1; // 0: custom APIs will run before the default APIs, 1: after, -1: reserved to default APIs
   parsingFunction?: TExplorerParsingFunction;
@@ -39,3 +61,7 @@ export interface ExplorerAPI {
   // defaults to 'eth'
   chainType?: 'btc' | 'evm' | 'eth';
 }
+
+export declare async function lookForTx (
+  { transactionId, chain, explorerAPIs = [] }: { transactionId: string; chain: SupportedChains; explorerAPIs?: ExplorerAPI[] }
+): Promise<TransactionData>;
