--- typings/explorers.ts
@@ -1,3 +1,0 @@
-export { TransactionData } from '../src/models/TransactionData';
-export { SupportedChains } from '../src/constants/blockchains';
-export { ExplorerURLs, ExplorerAPI, TExplorerFunctionsArray, IParsingFunctionAPI, TExplorerParsingFunction } from '../src/models/Explorers';
