--- typings/index.d.ts
@@ -1,2 +1,0 @@
-export * from './explorers';
-export * from './lookForTx';
